requirements.txt
mysql-connector-python==9.3.0

fastapi[standard]
uvicorn
pydantic
